Move home to its folder inside screens/
