+++ lib/screens/wrapper.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:innopolis_feedback/main.dart';
+import 'package:innopolis_feedback/screens/home/home.dart';
 import 'package:innopolis_feedback/screens/authenticate/authenticate.dart';
 import 'package:provider/provider.dart';
